feat: Add installation script and update API for unused parameters
- Added `cmake_install.cmake` for installation configuration and file management.
- Updated `setHyperparameters` and `updateModel` methods to suppress unused parameter warnings using [[maybe_unused]].
- Removed redundant macro definitions and unused helper functions in `logistic_regression.cpp`.
- Refactored `NeuralNet::train` to keep `total_error` for debugging while removing unused variables.
- Fixed sign comparison warning in `trie.cpp` by casting depth to size_t.
- Made `utils.cpp` empty as all utility functions are now defined inline in `utils.h`.
- Created minimal test stub in `test_algorithms.cpp` for future algorithm tests.

+++ src/core/data_structures/trie.cpp
@@ -92,7 +92,8 @@ namespace ml
 
             bool Trie::removeHelper(TrieNode *node, const string &word, int depth)
             {
-                if (depth == word.size())
+                // v190.0: Cast to size_t to fix sign comparison warning
+                if (static_cast<size_t>(depth) == word.size())
                 {
                     // Word is found. Mark the end as false.
                     if (!node->isEndOfWord)
